requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
sqlalchemy==2.0.18
pydantic==2.0.2
alembic==1.11.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
spotipy==2.23.0
openai==0.27.8
python-dotenv==1.0.0
httpx==0.24.1
psycopg2-binary==2.9.6
jinja2==3.1.2
redis==4.6.0
pytest==7.4.0